Fix/Review 엔티티 rating 필드 타입 변경 Byte -> Integer (#171)

+++ src/main/java/com/jxx/xuni/review/domain/RatingHandler.java
@@ -10,9 +10,9 @@ import java.util.List;
 public class RatingHandler {
 
     public double calculateAvg(List<Review> reviews) {
-        List<Byte> ratings = reviews.stream().map(r -> r.receiveRating()).toList();
+        List<Integer> ratings = reviews.stream().map(r -> r.receiveRating()).toList();
         Integer totalRating = 0;
-        for (Byte rating : ratings) {
+        for (Integer rating : ratings) {
             totalRating += rating;
         }
